requirements.txt
numpy
pandas
scikit-learn
torch
faiss-cpu
tqdm
pyyaml
